tags/3.5/requirements.txt
alabaster==0.7.10
appdirs==1.4.3
Babel==2.5.3
cached-property==1.3.1
cairocffi==0.8.0
CairoSVG==2.1.3
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
cssselect==1.0.3
cssselect2==0.2.1
defusedxml==0.5.0
diff-match-patch==20121119
Django==1.11.12
django-admin-list-filter-dropdown==1.0.1
django-admin-rangefilter==0.3.1
django-admin-sortable2==0.6.19
django-appconf==1.0.2
django-autocomplete-light==3.2.10
django-ckeditor==5.4.0
django-compressor==2.2
django-debug-toolbar==1.9.1
django-docs==0.2.1
django-flat-theme==1.1.4
django-import-export==1.0.0
django-js-asset==1.0.0
django-jsoneditor==0.0.12
django-mail-templated==2.6.5
django-modeladmin-reorder==0.2
django-object-actions==1.0.0
django-phonenumber-field==1.3.0
django-slug-preview==1.0.2
django-tag-parser==3.1
django-timezone-field==2.0
dnspython==1.15.0
docutils==0.14
et-xmlfile==1.0.1
future==0.16.0
html5lib==1.0.1
httplib2==0.10.3
idna==2.6
imagesize==1.0.0
isodate==0.6.0
jdcal==1.3
Jinja2==2.10
lxml==4.1.1
MarkupSafe==1.0
mysqlclient==1.3.12
odfpy==1.3.6
olefile==0.45.1
openpyxl==2.5.0
packaging==17.1
phonenumberslite==8.8.10
Pillow==5.0.0
pockets==0.5.1
pyBarcode==0.8b1
pycparser==2.18
Pygments==2.2.0
pyparsing==2.2.0
PyPDF2==1.26.0
pypng==0.0.18
PyQRCode==1.2.1
python-dateutil==2.6.1
python-memcached==1.59
pytidylib==0.3.2
pytz==2018.3
pyup-django==0.4.0
PyYAML==3.12
rcssmin==1.0.6
reportlab==3.4.0
requests==2.18.4
requests-toolbelt==0.8.0
rjsmin==1.0.12
simplejson==3.13.2
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.2
sphinx-rtd-theme==0.2.4
sphinxcontrib-napoleon==0.6.1
sphinxcontrib-websupport==1.0.1
sqlparse==0.2.4
tablib==0.12.1
tinycss==0.4
tinycss2==0.6.1
unicodecsv==0.14.1
urllib3==1.22
uwsgitop==0.10
webencodings==0.5.1
xlrd==1.1.0
xlwt==1.3.0
xmltodict==0.11.0
zeep==2.5.0
